Fix tweeting, now using tweetkit gem
Twitter changed it's requirements and limitations for API use... we apparently still can post tweets for free, with strict limits, once we fixed things up on our twitter developer portal end.

But we need to use twitter v2 api and oauth to do it. Which is really a pain! We seem to have found an in-development gem that will do it. It's a bit creaky. Doing so.

Got to stop using the old unmaintained gem.